Don't rotate base plane.
Associate vertex direction with rotation. Dot product along the vertex direction determines which side of the plane passing through the origin and perpendicular to the vertex direction the cell is. Origin will lie on the plane and not be rotated. Origin is part of every base plane and is never rotated.